Use Long integers for OCLC workids. (pt 2)

+++ rdfToSolr/src/main/java/edu/cornell/library/integration/indexer/updates/IdentifyCurrentSolrRecords.java
@@ -262,10 +262,10 @@ public class IdentifyCurrentSolrRecords {
 			pstmt.setInt(1, oclcId);
 			ResultSet rs = pstmt.executeQuery();
 			while (rs.next()) {
-				int workid = rs.getInt(1);
+				long workid = rs.getLong(1);
 				insertStmt.setInt(1, bibid);
 				insertStmt.setInt(2, oclcId);
-				insertStmt.setInt(3, workid);
+				insertStmt.setLong(3, workid);
 				insertStmt.addBatch();
 			}
 		}
